Move Photon parsing to generic schema system
